pyproject.toml
[project]
name = "samisk-ocr"
version = "1.0.0"
authors = [
    {name = "Tita Enstad", email = "tita.enstad@nb.no"}
]

requires-python = ">=3.11"

dependencies = [
    "ordbilder@git+ssh://git@github.com/Sprakbanken/hugin-munin-ordbilder",
    "pytesseract>=0.3.10",
    "tqdm>=4.66.5",
    "levenshtein>=0.25.1",
    "pre-commit>=3.8.0",
    "ruff>=0.5.6",
    "jiwer>=3.0.4",
    "pandas>=2.2.2",
    "Pillow>=10.4.0",
    "plotly>=5.23.0",
    "datasets>=2.21.0",
    "nb-tokenizer>=0.1.0",
    "wordcloud>=1.9.3",
    "dash>=2.17.1",
    "transformers[torch]>=4.44.1",
    "mlflow>=2.15.1",
    "evaluate>=0.4.2",
    "pydantic-settings>=2.4.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.5.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["I", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["samisk_ocr"]
